Several new features
Some specific days will have random unique messages display on the splash text rotation where the Twitch stream info is displayed.

Added "random" and "exclude" features to the NSFW commands to get random tags of a smaller list or exclude tags before selecting a random one.

Now, if images are sent to the bot, they will appear in a new channel in the private server. These are for holidays about lewd shit and users can send lewd pics to the bot and I will be able to see them because it's now basically gonewild.
**All users who send images will receive a response about the user having to be 18 years old to send lewd kinds of images, even if the image sent isn't any sort of lewd. Obvious underage images will be deleted and the user who sent the images will be reported.**