Added less specific querying for NSFW
For example, querying for just "futa" will return "futanari" in both random, exclude, and searching for a single tag.

If the query matches multiple tags, it will use all returned tags for random and exclude, but pick a random one for the single tag search.
Example: random or exclude querying for "gif" will use all tags containing "gif", which are "randomhentaigif", "nekogif", "girlsologif", "feetgif", "pussywankgif", and "pussygif"
Solo-queing for "gif" will return a random one of the previously mentioned tags.