Added 2 new commands
x!ai – gives the user an invite link so they can add the AI bot to a server to play against.
x!botsbyxyvy – gives the user a link to Bots by Xyvy site and a small amount of information about what the site is for.

Also added a bunch of new splash phrases and touched-up some other stuffs.